requirements.txt
discord.py
tweepy
python-dotenv
boto
boto3